+++ packages/database/src/outbox/entities/event-queue.entity.ts
@@ -1,0 +1,15 @@
+import { OutboxEntity } from './outbox.entity.js';
+
+export class EventQueueEntity extends OutboxEntity {
+  // Event metaadata
+  version!: number;
+
+  // Data
+  payload!: {
+    before?: any;
+    after: any;
+  };
+
+  // Timestamps
+  processedAt?: Date;
+}