+++ packages/database/src/outbox/writer/event-outbox.writer.ts
@@ -1,0 +1,5 @@
+import { EventQueueEntity } from '../entities/event-queue.entity.js';
+import { EventQueueModel } from '../models/event-queue.model.js';
+import { OutboxWriter } from './outbox.writer.js';
+
+export class EventQueueWriter extends OutboxWriter<EventQueueModel, EventQueueEntity> {}