+++ packages/database/src/outbox/writer/job-outbox.writer.ts
@@ -1,0 +1,5 @@
+import { JobsOutboxEntity } from '../entities/jobs-outbox.entity.js';
+import { JobsOutboxModel } from '../models/jobs-outbox.model.js';
+import { OutboxWriter } from './outbox.writer.js';
+
+export class JobsOutboxWriter extends OutboxWriter<JobsOutboxModel, JobsOutboxEntity> {}